Allow gasnet to honor --docdir
This is a reworking of a pull request by Christoph Junghans of LANL.

GASNet began prior to autoconf-2.60, which is the first version of
autoconf to support a --docdir configure option and a corresponding
$(docdir) substitution in the Makefiles.  GASNet has always provided
its *own* "docdir = ${datadir}/doc/gasnet" in the Makefile's and
fragments it generates.  This results in ignoring any user provided
--docdir setting, even when autoconf-2.60 or newer is used.

This commit makes the GASNet-provided definition of docdir conditional
on autoconf *not* defining it.