mxm-conduit: Fix a deadlock in exit coordination
This commit fixes a deadlock in the exit-coordination logic which can
occur due to a spin-poll loop in an AM handler.  The deadlock occurs
when the hander runs from the AM send path (before sending) and then
stalls waiting for a Reply to the AM that has yet to be sent.

This fix operates by shifting the "stall" until after the AM send has
actually completed, and is able to maintain the message ordering that
the stall was intended to ensure.  Specifically:
  If the recipient of an EXIT_REQ has issued an EXIT_ROLE_REQ, then
  it will not send the EXIT_REP (to the MASTER) until after it has
  received the EXIT_ROLE_REP (from the ROOT).

It is my understanding that this is necessary to ensure that MXM is
not shutdown prematurely, because that would cause the ROOT to hang
sending the EXIT_ROLE_REP.