mxm-conduit: Revise exit flow
This commit ports the timed exitcode-reduction logic from ibv-conduit
to mxm-conduit.  This ensures swift exits for the common case of a
collective call to gasnet_exit(), while still having robust logic to
deal with non-collecive exits.

Additionally this commit revises the manner in which we enforce the
mxm-specific requirement that receivers may not shutdown the transport
prior to receiving all messages (otherwise the sender may get stuck
trying to progress the send).  The previous mechanism was some
additional (relative to the ibv-conduit exit code) messaging and
tracking of messages.  Now we simply use an out-of-band barrier to
ensure all messaging has completed, placed at the *end* of the exit
coordination (rather than at the start as in the previous logic).