pshm: fix performance regression on netbsd
The multiconf logic for the BUPC runtime recently began passing
--disable-pthreads instead of --disable-par when configuring for
clang-upc.  On NetBSD this had the unexpected impact of killing the
performance with PSHM.  As had already been noted in the comments, we
know "-pthread" is needed at link time for acceptible performance,
though we do not know why.

This commit restores the previous performance by always passing
"-pthread" to link with PSHM on NetBSD, even if GASNet is not
supporting pthreads at all.  NOTE that this is *only* the link option
and therefore does not introduce any pthread "stuff" into the
compilation step (for instance -D_REENTRANT or include pthread.h)