amudp: allow AMUDP_SPMDStartup(NULL,NULL,...)
This commit enables AMUDP_SPMDStartup() to accept NULL for the argc
and argv arguments.  Most significantly, these changes rework the
spawning logic in amudp to use the environment for the passing of the
master location to the workers, rather than the command line.

Per Dan Bonachea's suggestion this commit implements (and documents)
an environment variable to specify the full path to the "env" command
on the worker nodes, overriding the default assumption that "env" is
in $PATH.