ssh-spawner: allow NULL argc and argv
This commit rewrites significant portions of the ssh-spawner logic to
enable gasnet_init(NULL, NULL) to work.  Most significantly, the args
required are now passed through the environment rather than on the
command line.  Ironically, it is also now necessary to pass the entire
command line in the ssh command, rather than through the control
socket, to ensure that main() actually receives them.