ibv/mxm: avoid some attempts to init ssh-spawner
Prior to this commit we unconditionally tried to initialize the
ssh-spawner before any other (unless disabled at configure time).
This has been designed to be "safe", but is entirely pointless if
the GASNET_IB_SPAWNER env var is explicitly set to "mpi" or "pmi".

Thic commit changes the spawn logic in both ibv- and mxm-conduit
to attempt to init the ssh-spawner only if the GASNET_IB_SPAWNER
env var is "ssh" or is not set.