spawners: factor logic common to all spawners
This commit makes several cleanups to the conduit-specific spawner
scripts to ease future maintenance and allow for possible changes in
the future that would provide better integration between the spawners
and the conduits.

1) Ensure that all conduit-specific spawner scripts that may
eventually invoke gasnetrun_mpi.pl will set all of the environment
variables it may use (even if some are currently not used because they
may be specific to a platform not supported by a given conduit).  This
is done via an AC_SUBST_FILE() which contains the common settings.

2) The following variables used by ibv- and mxm-conduit's common perl
script have been renamed to start with GASNET_SPAWN:
  + GASNET_IB_SPAWNER       -> GASNET_SPAWNER
  + GASNET_IB_CONDUIT       -> GASNET_SPAWN_CONDUIT
  + GASNET_IB_BOOTSTRAP_MPI -> GASNET_SPAWN_HAVE_MPI

3) Ensure that all conduit-specific spawner scripts set GASNET_SPAWNER
and GASNET_SPAWN_CONDUIT.  While these variables are currently used
only in ganetrun_ibv.pl (and thus by ibv- and mxm-conduits), having
them set uniformly will allow for some possible future improvements to
spawner integration.

4) One additional environment variable rename was made as part of the
factorization:
  + ENVCMD -> GASNET_ENVCMD
Not only is use of the GASNET_ namespace a better practice in general,
but is also consistent with the work (currently on a branch) to make
amudp pass master information to workers in the environment rather
than on the commandline.

Note that none of the renamed environment variables appear in user
documentation - they are internal in nature.