trace/debug: system-specific code to find argv
If the client calls gasnet_init(NULL,NULL), we can still find the
original argc and argv on some systems.  This commit implements the
necessary system-specific mechanims to locate argv[] on many systems,
thus allowing for full initialization of the tracing and backtracing
facilities that would otherwise be impared.

For Linux (including Cray's CNL/CLE) and Cygwin we extract argv[] from
the /proc/self/cmdline psuedo-file.

On Solaris we use /proc/self/psinfo to find the argc and the address
in our memory of the orignal argv, both as setup by the OS when the
process was launched.  There is a small chance that the memory has
been overwritten, but that would represent a pretty poorly behaved
program.  Additionally, this information is considered trust-worthy
enough that the "pargs" program uses it.

On Darwin/MacOSX we use the KERN_PROCARGS2 sysctl() to get argv[].
Testing shows both 10.4/ppc and 10.8/x86 to work.

On FreeBSD, NetBSD, and OpenBSD we use sysctl() with minor variations
in the MIB values and format of the resulting data.  The variations
are dealt with in a single implementation via the preprocessor.  For
BSD systems which mount a /proc filesystem (though most don't do so by
default) we can also obtain the cmdline from /proc as a second option
if the default sysctl()-based mechanism is unavailable or fails.

I have found that on BG/Q the only available information is in
/proc/self/cmdline which provides "plain text" with quoting removed.
Therefore, one cannot distinguish between the following:
   ./a.out 1 2
   ./a.out '1 2'
   './a.out 1' 2
So, there is no support for finding argv[] on BG/Q.
However, since we have no backtrace support on BG/Q, the only lost
functionality is the argv0 in some messages and the command line in
trace files.