Implement GASNET_NULL_ARGV_OK
This commit adds preprocessor logic to define GASNET_NULL_ARGV_OK to
either 0 or 1, with 1 indicating to the client that the conduit is
able to accept gasnet_init(NULL,NULL).

For mpi-conduit and conduits which (may) use mpi-spawner we are now
performing a configure-time probe for MPI_VERSION.  This is necessary
so that we can continue *not* including mpi.h in client code.