amudp: fix unclean exits with non-zero exitcodes
Prior to this commit a call to AMUDP_SPMDExit resulted in *all*
processes exiting with the provided exit code.  However, that exit
code had already been sent to the master for use as the job's
exitcode.  As a result of the process exiting non-zero with SPAWNFN=S
one would see multiple messages (often one from each process) like
      connection to [host] failed.
and a SIGTERM on the master.  Both are due to
      " || ( echo \"connection to %s failed.\" ; kill %i ) "
at the end of the ssh command line.  The SIGTERM is a potentially
more serious issue than the message because it resulted in loss of
the actual exit code.

Now the call from AMUDP_SPMDExit() to AMUDP_SPMDShutdown() passes
zero, and the master sends zero in its command telling workers to
exit.  These changes eliminate the messages and SIGTERM, and
preserve the desired job exit code.