ssh-spawner: fix for abnormal exits
In the case of certain abnormal exits, the ssh-spawner will ssh to a
remote node to issue a kill() (from the application executable, not
the shell).  Prior to this commit the ssh command to do this did not
include the "wrapper" command, if any.  Since the most common use of
the wrapper is to set LD_LIBRARY_PATH as necessary to sucessfully load
the executable, the lack of the wrapper could have (and *did* on
Carver) prevent these kills from succeeding.