pshm: spawner-specific bootstrapSNodeBroadcast
This commit adds implementations if the "SNodeBroadcast" operation
used by gasneti_pshm_init() to the three conduit-independed spawners:
mpi, ssh and pmi.

The mpi-spawner implementation uses point-to-point operations, with
the assurance that they are all within the same compute node.

The ssh-spawner implementation uses the naive algorithm in terms of
the bootstrapExchange operation, since there are no point-to-point
operations in ssh-spawner.

On non-Cray system the pmi-spawner implementation now uses a single
KVS_Put per supernode, rather then one for every process.  However, we
assume that the Cray-specific PMI_Allgather() is "pretty good" and
don't attempt to replace it with "discrete" ops, but we do simplify
the operation slightly vs. the naive implementation.
TODO: verify our assumption that PMI_Allgather is actually better.