xlc: use gcc inline asm when mc_func is missing
This commit makes changes to preprocessor logic so that when xlc lacks
support for "#pragma mc_func" it will use gcc-style asm instead
(assuming that is available).  These changes are intended to favor
xlc's native supprot when both are present, as it is expected to be
less buggy.

An unexpected bonus is that Clang for PowerPC can now use the inline
asm version of the LIFO code that was previously only reachable by
GCC.