configure: better probe for format attribute
This commit changes the configure-time probe for the __format_
attribute as applied to function pointers to better match the way we
use it.  Specifically we now probe for use both with and without
"extern".  This will allow us to better work with recent xlc, which
only allows this attribute on the declaration (no "extern") of a
function pointer, but not its definition.