pshm: modify the "reference" SNodeBroadcast
Prior to this commit, the "reference" implementation of the super-node
scoped broadcast, used to bootstrap PSHM, was based on an "in place"
Exchange operation.  That was chosen because it was tolerant of the
possibility that "src" was NULL or invalid on non-root callers.
However, AMMPI's bootstrap Exchange calls MPI_Allgather() without
making any checks for the in-place case, which is erroneous by the MPI
standard (and at least one implementation noticed or I'd not be making
this change).

This commit replaces the in-place Exchange with one that just uses the
callers "src" as-is.  This safe because all callers use a valid 'src'.
This is a property that was not known for certain when the work on
SNodeBroadcast began.