configure: fix quoting in GASNET_FIND_HEADER
The GASNET_FIND_HEADER was "misbehaving" when the client had defined
GASNET_FIND_HEADER_CPP to a value with quotes surrounding values with
spaces.  Specifically the expansion could (depending on your shell) be
split on the spaces without regards to the quotes.

This commit uses
  result=`eval "$GASNET_FIND_HEADER_CPP ..." | ...`
instead of
  result=`$GASNET_FIND_HEADER_CPP ... | ...`
to ensure the command is properly parsed for quoted strings.

NOTE: We probably have other instances of command-from-valriable that
are lacking an "eval".  However, the monster values produced by
gcc_as_cc.pl for XLC seem to be the only case that has had a visible
impact.