smp-conduit: directly implement value put/get
Previously changes were made to smp-conduit to make the Extended API
implementation inlineable (a property that had been lost when PSHM was
introduced).  However, that did not include direct implementations of
the value-based put and get, which remained in terms of the reference
implementation which reduced them to bulk put and get operations.
Since smp-conduit is 100% loopback, that meant that they all became
calls to GASNETE_UNALIGNED_MEMCPY, and thus to the regular memcpy.

This commit implements the value-based Puts and the blocking value-
based Get directly, which means use of GASNETE_VALUE_ASSIGN and
GASNETE_VALUE_RETURN, respectively.

No effort has been made with respect to the non-blocking value Get.

This should result in smaller/faster code.  It also just happens
to avoid an optimization "issue" with xlc-13.1 for linux/ppc64le which
I was investigating when I discovered that value put/get were using
memcpy().