ofi-conduit: merge recent spawner updates
This commit applies to ofi-conduit the changes made to ibv- and mxm-
conduits in the feature/pshm_snodebcast, as well as the following:
  ea0b861 spawners: factor logic common to all spawners
  6243502 ibv/mxm: avoid some attempts to init ssh-spawner

Additionally, the file
  ofi-conduit/contrib/gasnetrun_ofi.pl
is removed in favor of using
  ibv-conduit/contrib/gasnetrun_ibv.pl
which became sufficiently conduit-neutral in ea0b861.

Finally, this commit adds the ident strings that the spawner perl
script relies on for identifying which spawners are compiled in.