pmi-spawner: auto-detect more PMI installations
This commit merges the PMI probe logic which was specific to
ofi-conduit with the conduit-independent PMI configure probe.

This should add support for yod (from portals4's runtime lib) and (in
theory) for mpich's mpiexec.  However, in the case of mpiexec, the
lack of a pmi.h in most (all?) mpich installations is an issue.