ofi-conduit: introduce/use gasnetc_paratomic_t
This commit introduces a type gasnetc_paratomic_t and some associated
operations with are true atomic operations only in a PAR build (but
not in a PARSYNC build as with the pre-existing gasneti_weakatomic_t).

Use of this type allows cleanup of the code which operates on the
variable 'pending_rdma' in gasnet_ofi.c.