teams: first major step toward sparse team rep.
This commit is the first step toward an implementation of teams which
does not contain a dense representation of the members.  The eventual
goal is to store info on just O(log(P)) members other than self.
Specifically the (self +/- 2^i) peers are enough to implement a wide
variety of scalable collective operations including the dissemination
pattern (for barrier, gather-all, all-to-all, reduce-all) and binomial
tree (for broadcast, scatter, gather, reduce and scan), among others.

This commit makes the first big changes toward the goal of removing
the dense (rel2act_map) array of team members:

+ Construct and store the set of nodes at (self + 2^i) within the rank
  space of the team's nodes and its supernodes (if applicable) in each
  team.  This is the NEW data to be used as the basis of the sparse
  representation, but we cannot yet remove the dense array in favor of
  just this new data.

+ The barrier implementations (and their init code in particular) have
  been revised to use information from the NEW data.  This results in
  a significant simplification in the way each barrier's init code
  does setup for the hierarchical pshm barrier, since the sets of
  peers they need to use are already preconstructed for BOTH the cases
  with and without the hierarchical pshm barrier.

There is no longer any use of GASNETE_COLL_REL2ACT() in any barrier
implementation.  However, there is still substantial work required to
implement the full range of collectives in terms of just the new data.