teams: fix bug in team split (scratch space)
This commit fixes a bug in which the array of collectives scratch
spaces was being constructed incorrectly.  Specifically the code was
not indexing into the array of all scractch spaces to pick out the
ones that are members of the new teams.

This appears to have gone unnoticed for approx 6 years due to the
rare use of teams with scratch-space based collectives algorithms.