team_split: revise previous commit
Along with fixing a bug, the previous commit introduced use of realloc
to control the memory consumption of team_split.

This commit makes an extra read pass over the array to count the
team members *before* allocation.  This should be preferreable to
making two realloc() calls which may copy the arrays.