Implement GASNETT_{HOT,COLD} function decorations
This commit implements (and documents) HOT and COLD function
decorations.  These will expand to "__attribute__((__hot__))" (or
"__cold__") only if a configure probe has determined the attribute is
supported by the compiler (or is ignored without warning).  We take the
same care we do with other attribute-based decorations to treat the
configure-probed CC, MPI_CC and CXX independently, and to "squash" the
expansion when the compiler in use is none of those three.

As we go forward we can implement variants for other compliers,
perhaps by adding pragma-based versions as we have done with
GASNETT_{PURE,CONST}P.

These are being implemented to improve the portability of psm-conduit,
which has been written using the bare attributes.