Avoid SEGV on gasneti_realloc(ptr, 0)
This commit ensures that gasneti_realloc will *not* call
gasneti_memcheck(NULL) when called for zero size.  This is only a
concern in a build with debugging-malloc support enabled.

This showed up on a run of ibv-conduit on a system with no IB ports
configured when trying to realloc the port table from its max size
down to its actual size.