amudp: fix issues with quoting of arguments
Commands launched with GASNET_SPAWNFN=S or =C are evaluated TWICE:
once by the system() command, and again by a second shell (local
in the =C case or remote in the =S case).  This commit adds code to
protect single-quotes and the '$' and '`' characters that were
previously leading to errors.