gemini and aries conduits: support for hybrid apps
This commit adds code for manipulating the gni cookie value to ensure
that the MPI and GASNet libs in a hybrid application get distinct
value.  A user can build GASNet after enabling one of the two
variants, or add the "increment" variant into their own code between
the initializations of the two libs.

By default we have enabled the post-increment variant because it
should always be "harmless", and when GASNet initialization comes
before MPI initialization it fixes the problem.