Cray XE/XC: make "hugetlbfs" the default PSHM type
By using the "hugetlbfs" implementation of PSHM instead of XPMEM we
can avoid (at least) two known short-commings of XPMEM.

The first is documented as bug #3178 and involves the inability of the
GNI libraries to register memory mapped in using XPMEM, which is *not*
a problem with using mmap() on hugetlbfs pseudo-files.

The second is that upon process exit any XPMEM exports are immediately
revoked.  This has been seen to lead to a SIGBUS on other processes
which are accesing shared memory, when one process crashes.  The extra
crashes (typically 23 of them) that result made identifying the root
cause far more difficult (especially if the output gets interleaved
such that one or more SIGBUS failures actually appear *before* the
true crash).