amudp: fix issues with quoting of directory
This commit extendens the recent work to properly quote the command
line arguments within amudp to do the same with the "cd" command used
in the ssh command line.

Together with the previous fix for the command line args, this
resolves bug #938.

Additionally the code should be better prepared to deal with very
long directory names (without overflowing buffers).