PMI support: revise configure logic
This commit revises the configure logic that deals with PMI.

The main change is that HAVE_BOOTSTRAP_PMI will now be set on Cray
systems, which will enable pmi-based spawn of ofi-conduit jobs via
aprun.  Some additional factorization/cleanup was also performed,
which was enabled because the Cray-PMI case became less distinct.