Fix {GNI,IB}DISSEM barriers in non-default config
For the case that PSHM is enabled but GASNETI_PSHM_BARRIER_HIER has
been defined to zero, two conduit-specific implementations of the
RDMA-based dissemination barrier were crashing at runtime.  In both
cases the issue was that without the heirarchical shared memory
barrier the dissemination barrier must perform Puts to processes in
the same supernode.  Prior to this commit, that was attempted using
the network-based Put code, which failed because these conduits do not
create network connections to such peers.

This commit adds checks in the IBV and GNI conduits' "barrier_send"
code for the case of on-supernode destination, and converts these
to Stores instead of Puts.  The code is guarded by preprocessor checks
that PSHM is enabled and the hierarchical barrier is disabled.

Thanks to Yili Zheng for reporting the aries-conduit failures in
this non-default build configuration.