gemini/aries: update names to remove 'smsg'
The implementation of Active Message for GNI has not used the SMSG
APIs for a while now.  This commit updates names of variables and
functions, as well as a few comments, to eliminate references to
SMSG (which could cause confusion).

There are no "real" changes (to behavior or semantics) in this commit.