pshm barrier: implement GASNET_PSHM_BARRIER_HIER
This commit implements environment variable GASNET_PSHM_BARRIER_HIER
as a mechanism to disable, at run-time, the use of the hierarchical
shared-memory barrier.

This feature is in response to problems seen by the UPC++ team because
the shared-memory barrier code incudes a *blocking* implementation of
notify (in violation of the GASNet spec).  Longer-term, it is probably
beneficial to include a non-blocking notify (perhaps optionally) as an
alternative to disabling the shared-memory barrier entirely.