aries: initial implementation of local firehose
This commit contains an initial proof-of-concept implementation of
local-only firehose for Aries conduit.  This is disabled by default,
but one can pass --enable-gni-firehose at configure time to enable.

This code cannot be enabled for gemini-conduit, where it is almost
certain to fail because Gemini hardware lacks sufficient registration
resources.

This work is being pursued based on information from Cray which
confirms Khaled's experimental findings that the registration
capabilites of an Aries NIC are (for all practical purposes)
unbounded.  It is also motivated by Cray having confirmed that in the
multi-domain case we do not need to register memory with every domain,
because memory registration is global across Cdm handles created with
the same (ptag,cookie) pair.

Simple benchmarking ("testlarge -out") vs. the non-caching dynamic
registration inherited from the orginal (pre-Aries) gemini-conduit
shows significant latency reduction and bandwidth increase in the
middle range of payload sizes.

Debug runs (w/ and w/o pthreads) of the BUPC testsuite show no errors.