firehose: aggressive removal of covered regions
This commit allows a conduit to elect more aggressive removal of
"covered" regions from the Victim FIFO.  These are regions which
currently have no references, and cannot currently receive any
references because for every page within the region there is at least
one "better" region (one with greater length and/or forward extent).

Regions which become "covered" are placed at the dequeuing end of the
Victim FIFO (thus using it as LIFO in this case) to make it more
likely that they would be consumed (in preference to regions which
were still "best" for one or more of pages) when regions must be
evicted.  Since they are "useless", the code also used the opportunity
at each registration (cache miss) to deregister any covered regions.

Prior to this commit the opportunistic deregistration of covered
regions searched the FIFO for covered regions only until a non-covered
region was encountered.  With this commit a conduit can now choose (at
compile time) to have that search traverse the entire FIFO.