extended: fix !GASNETI_THREADINFO_OPT
This commit fixes a bug which preventing disabling the thread-info
optimization (GASNETI_THREADINFO_OPT).

GASNETI_THREADINFO_OPT will be defined to 1 by gasnet_help.h if it has
not been defined previously.  The intent is that a conduit (or CC
setting) could set it to 0 to disable.  However, prior to this commit
we were using "#ifdef GASNETI_THREADINFO_OPT" in three places, and
thus had no way to completely disable this feature.  Worse, since
there *were* testing "#if GASNETI_THREADINFO_OPT" in gasnet_help.h, a
build with GASNETI_THREADINFO_OPT=0 would get a SIGBUS or SIGSEGV at
runtime.