Add threadinfo-opt to link-time "idiot check"
Now that an end user can turn the threadinfo passing optimization ON or
OFF at configure time, and since that setting can chance the arguments
to many functions, we want to detect attempts to link to a library with
a different setting.  This commit add the bits necessary to make that
check.