aries: proof-of-concept uint64_t fetch-and-op
This commit implements fetch-and-op on 64-bit unsigned integers.  For
each suported operation there are five functions.

Three memory-to-memory APIs, modelled after non-bulk Gets:
  void            gasnetX_fetch<OP>_u64(dest, node, src, arg)
  gasnet_handle_t gasnetX_fetch<OP>_u64_nb(dest, node, src, arg)
  void            gasnetX_fetch<OP>_u64_nbi(dest, node, src, arg)
where the _nbi operation syncs with Gets.

Blocking and non-blocking memory-to-register APIs, which are modelled
after gasnet_get_val() and gasnet_get_nb_val():
  uint64_t               gasnetX_fetch<OP>_u64_val(node, src, arg)
  gasnet_valget_handle_t gasnetX_fetch<OP>_u64_nb_val(node, src, arg)

The "<OP>" is one of "add", "and", "or", "xor".

The "X" in "gasnetX_" is literal, and is intended to designate these
as eXtensions.

Since these are *typed* interfaces, the src and dest arguments are
pointers to uint64_t (rather than to void).