Configure: allow TLS on Sun compilers on x86
Prior to this commit we disabled use of __thread on all versions of
the Sun (aka Studio) compilers on x86-family systems (both Solaris and
Linux).  That was done due to compiler crashes with a now-ancient
version of the compiler (which no longer occurs with any version since
early 2009).  So, now we probe for support without regard to the
compiler faimily or target.  If somebody needs to use the old/broken
compiler we *do* provide --disable-tls.