Default disable threadinfo-opt on many platforms
Testing shows that for the major OSes on x86-64 and PPC systems the
use of our "clever" (but messy) threadinfo passing optimization is
generating code that is both larger and slower than use of __thread
and not passing the extra threadinfo argument (because the compiler
and linker do all the hard parts for us).

So, this commit changes the default for GASNETI_THREADINFO_OPT to 0 on
the platforms where this has been tested.

NOTE: One unfortunate risk of this change is that code may innocently
get committed which is broken when GASNETI_THREADINFO_OPT is enabled,
simply because the developer has tested only on systems where it is OFF
by default (which is all of the most common platforms).  This can (but
only partly) be addressed in the nightly testers.