aries/gemini multi-domain: change to polling mask
With this commit a value of 0 for GASNET_AM_DOMAIN_POLL_MASK will
result in every thread polling the AM queue on every poll call
(something that was not possible to express previously).

The frequency with all other values have been doubled, so that 1 is
now the value to poll AMs on 50% of poll calls by non-first domains
(where that was the meaning of 0 previously).

The default has changed from 127 to 255, so that the default polling
frequency as not actually changed.