aries/gemini: special-case small bounced gets
This commit separates the fully general post-completion copy operation
from the special case of aligned copies from the "immediate" buffer in
the gpd structure.  This makes for a slightly simpler code path for
the case of small out-of-segment Gets.  Additionally this addition
clears the way for some planned future additions.