smp-conduit: non-blocking notify in PSHM barrier
This commit modifies the barrier logic so that the default
shared-memory barrier in smp-conduit will no longer block in
gasnet_barrier_notify().

Some follow-up work is still required:

1) Eliminate blocking in notify for the hierarchical case where the
shared-memory barrier is combined with a network barrier.  This
requires a single new "pattern" be developed and applied to each of
the six PSHM-aware barrier implementations (3 conduit-independent
and 3 conduit-specific).

2) The kick function should be registered as a progress function in
SEQ builds of smp-conduit.

3) The kick function should be made thread-safe (in particular with
respect to its interactions with concurrent Notify, Try or Wait
calls).  This is required if a progress function is to be used in a
PAR build.