BG/Q: tolerate BG_MAPCOMMONHEAP != 1
In the default configuration for BG/Q, we use the "global heap" for
implementing PSHM.  However, that has the unfortunate dependence on
setting of BG_MAPCOMMONHEAP=1 in the job environment.  We take care to
set that variable in our spawner wrapper.  However, prior to this commit
if the user launched their job directly, without setting this env var,
their job would crash whenever running multiple processes per node
*unless* one instead set GASNET_SUPERNODE_MAXSIZE=1 to request that
GASNet not use PSHM.  This undesired behavior was true of all supported
conduits (smp, pami and mpi).

This commit adds logic to detect whether BG_MAPCOMMONHEAP is enabled in
the environment, and if not we behave as if GASNET_SUPERNODE_MAXSIZE is
set to 1. If it was not *actually* set to 1, then we also issue a
warning message.