amudp: honor GASNET_FS_SYNC (with default NO)
As was observed 6 years ago (see bug 2627), calling sync() when using
a cluster filesystem is a bad thing, resulting in measurable delays of
tens of seconds (and sometime complaints from sysadmins!!).

That issue was resolved by implementing a GASNET_FS_SYNC env var (with
a default of FALSE) to determine if gasnett_flush_streams() should
call sync().

This commit extends that same logic to the AMUDP equivalent call to
flush outstanding I/O.  This makes a *huge* difference in running
times of small benchmarks over udp-conduit, because the stream
flushing is invoked at each AMUDP_SPMDBarrier(), of which even a
minimal udp-conduit code typically has four.

Both GASNET_FS_SYNC and AMUDP_FS_SYNC are honored.