asm: fix SPARC v9 detection under Linux
Since removing support for SPARC CPUs prior to v9, we apparently have
not tested on Linux/SPARC.  Additionally, it appears that we may never
have actually utilized the SPARC v9 atomics under Linux.

This commit restores the ability to build GASNet on Linux/SPARC by
adding the proper preprocessor symbols to detect the CPU type which
gcc is targetting.