PSHM: fix Linux/SPARC support
We apparently have not tested Linux/SPARC since PSHM was added.  If we
had, then we would have discovered that the mmap() with MAP_SHARED has
special alignment constraints (due to cache architecture).  This was
causing EINVAL errno values from our mmap() calls with MAP_FIXED
whenever the requested segment size was not a multiple of the proper
alignment.  (The alignment is 16K vs. the page size of 8K).

This commit repairs PSHM support for Linux/SPARC by ensuring that we
use MAP_FIXED only with addresses resulting from "trimming" from the
*top* of the O/S-provided mappings.  This ensures we keep the correct
alignment (without the need to know what it is), just as we have done
for mappings on HUGETLBFS.