Cray XE/XC: return to XPMEM as default PSHM type
This reverts commit 596660d9805c810b8bf922110412d3b3afc32b7c.

Based on experiments by Weiqun Zhang it has been determined that use
of hugetlbfs for PSHM on the Cray systems can cause leaks of the
hugepage files.  Specifically, this has been observed when one or more
processes exits early due to insufficient hugepages (due to
fragmentation by earlier runs in the same batch job).

In the long term it would be beneficial to pursue use of either
fd-passing (SCM_RIGHTS) or mmap() of /proc/<pid>/fd/<N> as the
mechanism for sharing of hugepage files obtained from calls to
hugetlbfs_unlinked_fd().  This would give the benefits (no SIGBUS in
other processes when a single process exits) without the possibility
of a leak.

Additionally, we should pursue fixing the leak which is probably also
present with --enable-pshm-file on OSes w/o POSIX or SysV shared
memory support.