segment: remove GASNETI_SEGMENT_DISALIGN_BIAS code
This commit removes the "GASNETI_SEGMENT_DISALIGN_BIAS" debugging
code.  This code was added when segment alignment became optional, and
was intended to ensure all conduits and clients were able to deal
correctly with unaligned segments.  For a significant time now the
unaligned case has been the more common one, and the likelihood of a
client or conduit NOT correctly supporting this case is therefore very
low.  Additionally, exceptions for things such as hugepage alignment
have reduced the cases in which it is possible to apply the bias at
all.  Finally, the remaining cases in which the bias can be applied
have become so uncommon that testing any changes that might interact
with the bias has become quite difficult.