aries: fix firehose with respect to bug 3178
The (local-only) firehose support in aries-conduit was undertaken
under the assumption that we were switching from XPMEM to HUGETLBFS as
the default PSHM implementation.  However, that change was reverted
due to problems seen in testing.  So, this commit is to provide a
work-around for the "bug 3178" problem: GNI cannot register memory
which was imported via XPMEM.  Additionally, we address a recently
discovered related problem with the inability to register read-only
pages.  (Note that the read-only problem needs to be dealt with
independent of the XPMEM vs HUGETLBS choice, making the work to deal
with the XPMEM-specific issue almost "free").

The approach when faced with a failing GNI registration call is to
create a firehose registration which essentially caches the inability
to register the memory.  Meanwhile the code which makes registration
requests from firehose now checks the result to distinguish the case
of a failed registration (and fall back to bounce buffers).  The
layout of a GNI memory handle is such that we can easily construct
an invalid one and distinguish it from a valid one.

We are preventing firehose from merging normal and failed memory
registrations through the addition of a new hook that allows conduits
to prevent firehose from merging particular regions.