ibv-conduit: move lid table into shared memory
This commit leverages PSHM support to keep just a single copy of the
global lid table per supernode.

The lid table is 2-bytes per node.  So, for a 4K-core run consisting
of 256 nodes of 16 cores each, this means one copy (8K) instead of 16
(128K).  So, this is not a huge savings, but this is the first step to
reduce the memory footprint by moving common data into shared memory.