ssh-spawner: reduce (again) comms for SNodeBcast
This commit replaces a size_t (giving a length) in the communications
protocol for the supernode-scoped broadcast operation, with a
gasnet_node_t (giving a node count).  For the case of ibv-conduit this
is a reduction from 8 bytes to 2, which is significant since the
payload which follows may be as small as 2 bytes (at a typical leaf in
the tree).